Fix hidden overloaded virtual function in BitParameter.h
accessAsInteger function prototype is defined in BaseParameter
abstract class with an uint32_t type. Nevertheless in BitParameter
the overloaded function take an uin64_t. It seems that GCC accepts
wrongly this mistake but clang don't.
This patch corrects the function prototype accordingly to the one
defined in BaseParameter. It allows the project to be compiled
with clang.

+++ parameter/BitParameterType.h
@@ -53,7 +53,7 @@ public:
     void fromBlackboard(std::string& strValue, const uint64_t& uiValue, CParameterAccessContext& parameterAccessContext) const;
     // Integer
     bool toBlackboard(uint64_t uiUserValue, uint64_t& uiValue, CParameterAccessContext& parameterAccessContext) const;
-    void fromBlackboard(uint64_t& uiUserValue, uint64_t uiValue, CParameterAccessContext& parameterAccessContext) const;
+    void fromBlackboard(uint32_t& uiUserValue, uint64_t uiValue, CParameterAccessContext& parameterAccessContext) const;
     // Access from area configuration
     uint64_t merge(uint64_t uiOriginData, uint64_t uiNewData) const;
